+++ internal/init.go
@@ -1,0 +1,2 @@
+//Package internal implements few hacks to allow pomerium embeddeding
+package internal
