+++ pomerium/envoy/envoy_darwin_amd64.go
@@ -11,6 +11,5 @@ var rawBinary []byte
 //go:embed bin/envoy-darwin-amd64.sha256
 var rawChecksum string
 
-const (
-	enabled = true
-)
+//go:embed bin/envoy-darwin-amd64.version
+var rawVersion string
