+++ pomerium/envoy/validate.go
@@ -1,0 +1,7 @@
+package envoy
+
+// A ValidateResult is the result of validation.
+type ValidateResult struct {
+	Valid   bool
+	Message string
+}
