+++ scripts/get-envoy.bash
@@ -4,7 +4,7 @@ set -euo pipefail
 PATH="$PATH:$(go env GOPATH)/bin"
 export PATH
 
-_envoy_version=1.20.1
+_envoy_version=1.21.3
 _dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)/../pomerium/envoy/bin"
 _target="${TARGET:-"$(go env GOOS)-$(go env GOARCH)"}"
 _envoy_binary_name="envoy-$_target"
