+++ tools.go
@@ -5,6 +5,7 @@ package pomerium
 
 import (
 	_ "github.com/client9/misspell/cmd/misspell"
+	_ "github.com/golang/mock/mockgen"
 	_ "github.com/golangci/golangci-lint/cmd/golangci-lint"
 	_ "sigs.k8s.io/controller-runtime/tools/setup-envtest"
 	_ "sigs.k8s.io/controller-tools/cmd/controller-gen"
